file_read_line: Return in case of error to avoid returning undef
A lone return statement is equivalent to `return undef` (or returning an
undefined variable) in scalar context. Obviously, both evaluate to the
same: false. But in list context, a lone return statement returns an
empty list, while `return undef` returns a list with one element: undef.
The first is considered false, while the one element list is considered
true.

This change is just hardening, no (current) code paths were susceptible
to this problem.